LPS-55414 Extract user bag creation logic into UserBagFactory, rename test accordingly
