+++ portal-service/src/com/liferay/portal/security/permission/UserBagFactoryUtil.java
@@ -1,0 +1,59 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.portal.security.permission;
+
+import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.security.pacl.permission.PortalRuntimePermission;
+import com.liferay.registry.Registry;
+import com.liferay.registry.RegistryUtil;
+import com.liferay.registry.ServiceTracker;
+
+/**
+ * @author Preston Crary
+ */
+public class UserBagFactoryUtil {
+
+	public static UserPermissionCheckerBag create(long userId)
+		throws PortalException {
+
+		return getUserBagFactory().create(userId);
+	}
+
+	public static PermissionCheckerBag create(long userId, long groupId)
+		throws PortalException {
+
+		return getUserBagFactory().create(userId, groupId);
+	}
+
+	public static UserBagFactory getUserBagFactory() {
+		PortalRuntimePermission.checkGetBeanProperty(UserBagFactoryUtil.class);
+
+		return _instance._serviceTracker.getService();
+	}
+
+	private UserBagFactoryUtil() {
+		Registry registry = RegistryUtil.getRegistry();
+
+		_serviceTracker = registry.trackServices(UserBagFactory.class);
+
+		_serviceTracker.open();
+	}
+
+	private static final UserBagFactoryUtil _instance =
+		new UserBagFactoryUtil();
+
+	private final ServiceTracker<?, UserBagFactory> _serviceTracker;
+
+}
